green-evaluator/requirements.txt
# Green GAIA Agent Requirements
# Generated from pyproject.toml for convenience
# Install with: pip install -r requirements.txt

# Core dependencies
a2a-sdk>=0.3.5
httpx>=0.28.0
pydantic>=2.11.9
python-dotenv>=1.1.1

# LLM and Agent capabilities
google-genai>=1.0.0
google-adk>=1.14.1

# Development dependencies
pytest>=8.3.2
pytest-asyncio>=0.23.7
pytest-cov>=4.1.0

# Linting and type checking
ruff>=0.4.6
mypy>=1.15.0
types-requests>=2.32.0
